requirements.txt
python-dotenv
openai
google-genai

pandas 
numpy
scikit-learn

torch
transformers

huggingface-hub

langchain
langchain-core

langchain-openai
langchain-google-genai
langchain-anthropic
langchain-huggingface

streamlit

